+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-## v0.1.0
+## v0.2.0 (2024-01-16)
+
+- Add `chat/completions` endpoint.
+- Improve ExDocs for developer.
+
+## v0.1.0 (2024-01-15)
 
 - Add CHANGELOG.md
 - Add `models` endpoint
 - Add `Config` overwrite feature
 - Add ex_docs
-- Add `Dependabot` and `c.yml`
+- Add `Dependabot` and `ci.yml`